Revert to 'permissive' term
On advice, I'm going to retain the term permissive as it has been pointed out that both GPL and MIT require attribution so would be incorrect to simply say "attribution" is a license style. I've removed "academic" as the license style as this could be confused with that actual licence named the "Academic Free Licence" i.e. https://opensource.org/licenses/AFL-3.0.